% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_dubossarsky_network.R
\name{bipartite_to_unipartite_slow}
\alias{bipartite_to_unipartite_slow}
\title{Inefficient but more transparent implementation of w_ij calculation}
\usage{
bipartite_to_unipartite_slow(bpg)
}
\arguments{
\item{bpg}{Cue by response matrix}
}
\value{
Cue by cue matrix of w_ij
}
\description{
Inefficient but more transparent implementation of w_ij calculation
}
