% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_unweighted_network.R
\name{create_unweighted_network}
\alias{create_unweighted_network}
\title{Generate unweighted directed network from word associations}
\usage{
create_unweighted_network(cues, targets)
}
\arguments{
\item{cues}{A factor of character vector. Unique values determine network nodes}

\item{targets}{A factor of character vector. \code{cue -> target} associations determine network edges.}
}
\value{
A square adjacency matrix representing an unweighted directed
  network. Directed connections from row to column node.

For the associative networks, a pair of cues was connected by a directed link from
the cue word to the target word if that cue-target relationship was reported
in the association norms.
}
\description{
The number of cues and targets must be equal.
}
