% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_network.R
\name{save_network}
\alias{save_network}
\title{Uniform interface for saving networks}
\usage{
save_network(
  input,
  netlab,
  weighted,
  cond,
  chunkBy = NULL,
  nresp = NULL,
  netdir = NULL
)
}
\arguments{
\item{input}{The network or association data for composing the network.}

\item{netlab}{Network label (CHILDES, CoxHae, or SWOW).}

\item{weighted}{A logical (should a weighted nework be produced and written?)}

\item{cond}{A flexible variable that is used to differentiate the network
(and possibly subset associations).}

\item{chunkBy}{An informative labels, used with CHILDES networks, to indicate
whether windows were "chunked" by transcript or utterance. If chunked by
utterance, associations cannot be tracked across utterances. If chunked by
transcript, then associations can be tracked across utterances in the same
transcript, but not across transcripts.}

\item{nresp}{An informative label (and filter) for CoxHae and SWOW
associations. For each cue, a participant was asked to provide 3 word
associations. This variable indicates how many of those responses should be
considered when constructing the network.}

\item{netdir}{The basename for the directory where the network should be
written to. This will be a subfolder of the \code{project_root/network/}
directory.}
}
\description{
Uniform interface for saving networks
}
