% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_dubossarsky_network.R
\name{create_dubossarsky_net}
\alias{create_dubossarsky_net}
\title{Generate weighted directed network from word associations}
\usage{
create_dubossarsky_net(cues, targets)
}
\arguments{
\item{cues}{A factor of character vector. Unique values determine network nodes}

\item{targets}{A factor of character vector. \code{cue -> target} associations determine network edges.}
}
\value{
An igraph object representing weighted directed network.

Dubossarsky, De Deyne, and Hills (2017) project the cue by response
"bipartite graph" to a "unipartite graph".

Quoting from Dubossarsky et al. (2017, p1569):

"[The] edge weights are defined asymmetrically as the count of associates from
\eqn{\text{word}_i}{word_i} to the associates shared with
\eqn{\text{word}_j}{word_j}. This is normalized by the relative
distinctiveness of the shared associates by dividing the number of times
these associations are shared over all the cue words.

\deqn{w_{i,j} = \sum_{i=1}^P \frac{w_{i,p}}{N_p - 1}}

"\eqn{w_{i,j}} is an edge's weight in the one-mode graph directed from node i
to node j. P is the number of shared associations between cue words i and j.
w_{i,p} is the number of times cue i produced p as an associate. N_p is the
number of cue words that produced p. Cue words idiosyncratic responses were
removed prior to projection."

See \code{bipartite_to_unipartite} for implementation details.
}
\description{
Generate weighted directed network from word associations
}
\references{
Dubossarsky, H., De Deyne, S., & Hills, T. T. (2017). Quantifying the
structure of free association networks across the life span. Developmental
psychology, 53(8), 1560.
}
\seealso{
\code{\link[netbuildr]{bipartite_to_unipartite}}
}
